Laba-4-Threads
Develop a multi-threaded application.
Use of the package java.util.concurrent.
Do not use the word synchronized.
All entities wishing to access a resource must be streams.
Use OOP features.
Do not use the graphical interface. The application must be con-
solo.
1. Port. The ships come into port for unloading / loading of containers. Number
containers that are currently in the port and the ship must
be non-negative and exceeds a predetermined load sud-
and on the capacity of the port. The port has several berths. one
berth may be a ship. The ship can be loaded at the pier,
unloaded, or perform both actions.